Add workaround versioning macros
Add the `DPLX_XDEF_WORKAROUND` macro which enables in-code workarounds
for older but still supported (compiler) versions with an additional
grouping disable switch.

Add the `DPLX_XDEF_WORKAROUND_TESTED_AT` macro which ought to be used
for in-code workarounds where the required bugfix has not yet been
associated with an available version. Therefore we provide the
`flagOutdated` parameter which maps all occurences which should be
checked again for the current component version to compiler errors.
E.g. if a workaround was last validated for version 1.0.0 and we now
build with version 1.0.1, we would emit a compiler error if (and only
if) `flagOutdated=1`.

Add specializations of the above named `DPLX_CNCR_WORKAROUND` and
`DPLX_CNCR_WORKAROUND_TESTED_AT` which switch on
`DPLX_CNCR_DISABLE_WORKAROUNDS` and
`DPLX_CNCR_FLAG_OUTDATED_WORKAROUNDS`.