feat: Add `if consteval` polyfill macro
Add a macro which expands to `if consteval` in C++23 and falls back to
`if (std::is_constant_evaluated())` in C++20 mode. Obviously this
doesn't polyfill the immediate execution context, but it allows the
compiler to elide the compile time algorithm from the machine code if
the code is compiled with C++23.